Refactorización en carpeta utils. Unificación de constantes y reacomodado de archivos
