+++ src/main/kotlin/quickfix/utils/Constants.kt
@@ -1,4 +1,6 @@
 package quickfix.utils
 
+const val FRONTEND_URL = "http://localhost:5173"
 const val MAXIMUM_DEBT: Double = 1000.0
 const val PAGE_SIZE: Int = 4
+const val COMISSION = 0.4
