--- src/main/kotlin/quickfix/utils/Ports.kt
@@ -1,5 +1,0 @@
-package quickfix.utils
-
-const val FRONTEND_URL = "http://localhost:5173"
-const val CONFIRM_FRONTEND_URL = "/confirm"
-const val RECOVERY_FRONTEND_URL = "/confirm"
