--- src/main/kotlin/quickfix/utils/maxDebtAllowed.kt
@@ -1,4 +1,0 @@
-package quickfix.utils
-
-
-val  MAX_DEBT_ALLOWED = 50000.0
